Fix unintended transferred of print language from Sales Order to Purchase Order.
The Supplier Master has the needed print language therefore our script gets the print value from supplier doctype with the supplier name as the filter.
On save of the PO all line items item_language are set from the returned value of the Supplier master that is also set as the purchase order print language.